refactor flavors to use a collection instead of a store on the regional server collection
