+++ mimic/model/flavor_objects.py
@@ -1,279 +1,149 @@
 """
-Model objects for mimic flavors.
+Model objects for the flavors.
 """
 
-from characteristic import attributes
+from characteristic import attributes, Attribute
+from json import dumps
+from mimic.model.flavors import (
+    RackspaceStandardFlavor, RackspaceComputeFlavor, RackspaceMemoryFlavor,
+    RackspaceOnMetalFlavor, RackspaceIOFlavor, RackspaceGeneralFlavor,
+    RackspacePerformance1Flavor, RackspacePerformance2Flavor)
+
+from twisted.web.http import NOT_FOUND
 
 
-@attributes(['flavor_id', 'tenant_id', 'name', 'ram', 'vcpus', 'rxtx', 'disk'])
-class Flavor(object):
+@attributes(['nova_message'])
+class BadRequestError(Exception):
     """
-    A Flavor object
+    Error to be raised when bad input has been received to Nova.
     """
 
-    static_defaults = {
-        "swap": "",
-        "OS-FLV-EXT-DATA:ephemeral": 0,
+
+def _nova_error_message(msg_type, message, status_code, request):
+    """
+    Set the response code on the request, and return a JSON blob representing
+    a Nova error body, in the format Nova returns error messages.
+
+    :param str msg_type: What type of error this is - something like
+        "badRequest" or "itemNotFound" for Nova.
+    :param str message: The message to include in the body.
+    :param int status_code: The status code to set
+    :param request: the request to set the status code on
+
+    :return: dictionary representing the error body
+    """
+    request.setResponseCode(status_code)
+    return {
+        msg_type: {
+            "message": message,
+            "code": status_code
+        }
     }
 
-    def links_json(self, absolutize_url):
-        """
-        Create a JSON-serializable data structure describing the links to this
-        flavor.
-        """
-        return [
-            {
-                "href": absolutize_url("v2/{0}/flavors/{1}"
-                                       .format(self.tenant_id, self.flavor_id)),
-                "rel": "self"
-            },
-            {
-                "href": absolutize_url("{0}/flavors/{1}"
-                                       .format(self.tenant_id, self.flavor_id)),
-                "rel": "bookmark"
-            }
-        ]
 
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a flavor.
-        """
-        return {
-            "quota_resources": "instances=mimic-instances,ram=mimic-ram",
-        }
-
-    def brief_json(self, absolutize_url):
-        """
-        Brief JSON-serializable version of this flavor, for the non-details
-        list flavors request.
-        """
-        return {
-            "id": self.flavor_id,
-            "links": self.links_json(absolutize_url),
-            "name": self.name
-        }
-
-    def detailed_json(self, absolutize_url):
-        """
-        Long-form JSON-serializable object representation of this flavor, as
-        returned by either a GET on this individual flavor or a member in the
-        list returned by the list-details request.
-        """
-        template = self.static_defaults.copy()
-        template.update({
-            "id": self.flavor_id,
-            "links": self.links_json(absolutize_url),
-            "name": self.name,
-            "ram": self.ram,
-            "vcpus": self.vcpus,
-            "rxtx_factor": self.rxtx,
-            "disk": self.disk,
-            "OS-FLV-WITH-EXT-SPECS:extra_specs": self.extra_specs_json()
-        })
-        return template
-
-
-class RackspaceStandardFlavor(Flavor):
+def not_found(message, request):
     """
-    A Rackspace standard flavor object representation
+    Return a 404 error body associated with a Nova not found error.
+    Also sets the response code on the request.
+
+    :param str message: The message to include in the bad request body.
+    :param request: The request on which to set the response code.
+
+    :return: dictionary representing the error body.
     """
-    flavors = {"512MB Standard Instance": {"id": "2", "ram": 512, "vcpus": 1, "rxtx_factor": 80,
-                                           "disk": 20},
-               "1GB Standard Instance": {"id": "3", "ram": 1024, "vcpus": 1, "rxtx_factor": 120,
-                                         "disk": 40},
-               "2GB Standard Instance": {"id": "4", "ram": 2048, "vcpus": 2, "rxtx_factor": 240,
-                                         "disk": 80},
-               "4GB Standard Instance": {"id": "5", "ram": 4096, "vcpus": 2, "rxtx_factor": 400,
-                                         "disk": 160},
-               "8GB Standard Instance": {"id": "6", "ram": 8192, "vcpus": 4, "rxtx_factor": 600,
-                                         "disk": 320},
-               "15GB Standard Instance": {"id": "7", "ram": 15360, "vcpus": 6, "rxtx_factor": 800,
-                                          "disk": 620},
-               "30GB Standard Instance": {"id": "8", "ram": 30720, "vcpus": 8, "rxtx_factor": 1200,
-                                          "disk": 1200}}
-
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a standard flavor.
-        """
-        return {
-            "class": "standard1",
-            "policy_class": "standard_flavor"
-        }
+    return _nova_error_message("itemNotFound", message, NOT_FOUND, request)
 
 
-class RackspaceComputeFlavor(Flavor):
+@attributes(
+    ["tenant_id", "region_name", "clock",
+     Attribute("flavors_store", default_factory=list)]
+)
+class RegionalFlavorCollection(object):
     """
-    A Rackspace compute flavor object representatione
+    A collection of flavors, in a given region, for a given tenant.
     """
-    flavors = {"3.75 GB Compute v1": {"id": "compute1-4", "ram": 3840, "vcpus": 2, "rxtx_factor": 312.5,
-                                      "disk": 0},
-               "7.5 GB Compute v1": {"id": "compute1-8", "ram": 7680, "vcpus": 4, "rxtx_factor": 625,
-                                     "disk": 0},
-               "15 GB Compute v1": {"id": "compute1-15", "ram": 15360, "vcpus": 8, "rxtx_factor": 1250,
-                                    "disk": 0},
-               "30 GB Compute v1": {"id": "compute1-30", "ram": 30720, "vcpus": 16, "rxtx_factor": 2500,
-                                    "disk": 0},
-               "60 GB Compute v1": {"id": "compute1-60", "ram": 61440, "vcpus": 32, "rxtx_factor": 5000,
-                                    "disk": 0}}
-
-    def extra_specs_json(self):
+    def flavor_by_id(self, flavor_id):
         """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a compute flavor.
+        Retrieve a :obj:`Flavor` object by its ID.
         """
-        return {
-            "class": "compute1",
-            "policy_class": "compute_flavor"
-        }
+        if len(self.flavors_store) < 1:
+            self.create_flavors_list()
+        for flavor in self.flavors_store:
+            if flavor.flavor_id == flavor_id:
+                return flavor
+
+    def create_flavors_list(self):
+        """
+        Generates the data for each flavor in each flavor class
+        """
+        rackspace_flavors = [RackspaceStandardFlavor, RackspaceComputeFlavor,
+                             RackspacePerformance1Flavor, RackspaceOnMetalFlavor,
+                             RackspacePerformance2Flavor, RackspaceMemoryFlavor,
+                             RackspaceIOFlavor, RackspaceGeneralFlavor]
+
+        if len(self.flavors_store) < 1:
+            for flavor_class in rackspace_flavors:
+                for flavor, flavor_spec in flavor_class.flavors.iteritems():
+                    flavor_name = flavor
+                    flavor_id = flavor_spec['id']
+                    ram = flavor_spec['ram']
+                    vcpus = flavor_spec['vcpus']
+                    network = flavor_spec['rxtx_factor']
+                    disk = flavor_spec['disk']
+                    tenant_id = self.tenant_id
+                    flavor = flavor_class(flavor_id=flavor_id, tenant_id=tenant_id,
+                                          name=flavor_name, ram=ram, vcpus=vcpus,
+                                          rxtx=network, disk=disk)
+                    self.flavors_store.append(flavor)
+
+    def list_flavors(self, include_details, absolutize_url):
+        """
+        Return a list of flavors with details.
+        """
+        if len(self.flavors_store) < 1:
+            self.create_flavors_list()
+        flavors = []
+        for flavor in self.flavors_store:
+            if self.region_name != "IAD" and isinstance(flavor, RackspaceOnMetalFlavor):
+                continue
+            if include_details:
+                flavors.append(flavor.detailed_json(absolutize_url))
+            else:
+                flavors.append(flavor.brief_json(absolutize_url))
+        result = {"flavors": flavors}
+
+        return dumps(result)
+
+    def get_flavor(self, http_get_request, flavor_id, absolutize_url):
+        """
+        Return a flavor object if one exists from the list `/flavors` api,
+        else creates and adds the flavor to the :obj: `flavors_store`.
+        If the `flavor_id` is listed in `mimic.canned_responses.mimic_presets`,
+        then will return 404.
+        """
+        flavor = self.flavor_by_id(flavor_id)
+        if flavor is None:
+            return dumps(not_found("The resource could not be found.",
+                                   http_get_request))
+        return dumps({"flavor": flavor.detailed_json(absolutize_url)})
 
 
-class RackspaceGeneralFlavor(Flavor):
+@attributes(["tenant_id", "clock",
+             Attribute("regional_collections", default_factory=dict)])
+class GlobalFlavorCollections(object):
     """
-    A Rackspace general flavor object representation
+    A :obj:`GlobalFlavorCollections` is a set of all the
+    :obj:`RegionalFlavorCollection` objects owned by a given tenant.  In other
+    words, all the flavor objects that a single tenant owns globally.
     """
-    flavors = {"1 GB General Purpose v1": {"id": "general1-1", "ram": 1024, "vcpus": 1,
-                                           "rxtx_factor": 200, "disk": 20},
-               "2 GB General Purpose v1": {"id": "general1-2", "ram": 2048, "vcpus": 2,
-                                           "rxtx_factor": 400, "disk": 40},
-               "4 GB General Purpose v1": {"id": "general1-4", "ram": 4096, "vcpus": 4,
-                                           "rxtx_factor": 800, "disk": 80},
-               "8 GB General Purpose v1": {"id": "general1-8", "ram": 8192, "vcpus": 8,
-                                           "rxtx_factor": 1600, "disk": 160}}
 
-    def extra_specs_json(self):
+    def collection_for_region(self, region_name):
         """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a general purpose flavor.
+        Get a :obj:`RegionalFlavorCollection` for the region identified by the
+        given name.
         """
-        return {
-            "class": "general1",
-            "policy_class": "general_flavor"
-        }
-
-
-class RackspaceIOFlavor(Flavor):
-    """
-    A Rackspace IO flavor object representation
-    """
-    flavors = {"15 GB I/O v1": {"id": "io1-15", "ram": 15360, "vcpus": 4, "rxtx_factor": 1250,
-                                "disk": 40},
-               "30 GB I/O v1": {"id": "io1-30", "ram": 30720, "vcpus": 8, "rxtx_factor": 2500,
-                                "disk": 40},
-               "60 GB I/O v1": {"id": "io1-60", "ram": 61440, "vcpus": 16, "rxtx_factor": 5000,
-                                "disk": 40},
-               "90 GB I/O v1": {"id": "io1-90", "ram": 92160, "vcpus": 24, "rxtx_factor": 7500,
-                                "disk": 40},
-               "120 GB I/O v1": {"id": "io1-120", "ram": 122880, "vcpus": 32, "rxtx_factor": 10000,
-                                 "disk": 40}}
-
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a memory flavor.
-        """
-        return {
-            "class": "io1",
-            "policy_class": "io_flavor"
-        }
-
-
-class RackspaceMemoryFlavor(Flavor):
-    """
-    A Rackspace memory flavor object representation
-    """
-    flavors = {"15 GB Memory v1": {"id": "memory1-15", "ram": 15360, "vcpus": 2, "rxtx_factor": 625,
-                                   "disk": 0},
-               "30 GB Memory v1": {"id": "memory1-30", "ram": 30720, "vcpus": 4, "rxtx_factor": 1250,
-                                   "disk": 0},
-               "60 GB Memory v1": {"id": "memory1-60", "ram": 61440, "vcpus": 8, "rxtx_factor": 2500,
-                                   "disk": 0},
-               "120 GB Memory v1": {"id": "memory1-120", "ram": 122880, "vcpus": 16, "rxtx_factor": 5000,
-                                    "disk": 0},
-               "240 GB Memory v1": {"id": "memory1-240", "ram": 245760, "vcpus": 32,
-                                    "rxtx_factor": 10000, "disk": 0}}
-
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a flavor.
-        """
-        return {
-            "class": "memory1",
-            "policy_class": "memory_flavor"
-        }
-
-
-class RackspaceOnMetalFlavor(Flavor):
-    """
-    A Rackspace onMetal flavor object representation
-    """
-    flavors = {"OnMetal Compute v1": {"id": "onmetal-compute1", "ram": 32768, "vcpus": 20,
-                                      "rxtx_factor": 10000, "disk": 32},
-               "OnMetal IO v1": {"id": "onmetal-io1", "ram": 131072, "vcpus": 40,
-                                 "rxtx_factor": 10000, "disk": 32},
-               "OnMetal Memory v1": {"id": "onmetal-memory1", "ram": 524288, "vcpus": 24,
-                                     "rxtx_factor": 10000, "disk": 32}}
-
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for an onMetal flavor.
-        """
-        return {
-            "quota_resources": "instances=onmetal-compute-v1-instances,ram=onmetal-compute-v1-ram",
-            "class": "onmetal",
-            "policy_class": "onmetal_flavor"
-        }
-
-
-class RackspacePerformance1Flavor(Flavor):
-    """
-    A Rackspace perfomance flavor object representation
-    """
-    flavors = {"1 GB Performance": {"id": "performance1-1", "ram": 1024, "vcpus": 1, "rxtx_factor": 200,
-                                    "disk": 20},
-               "2 GB Performance": {"id": "performance1-2", "ram": 2048, "vcpus": 2, "rxtx_factor": 400,
-                                    "disk": 40},
-               "4 GB Performance": {"id": "performance1-4", "ram": 4096, "vcpus": 4, "rxtx_factor": 800,
-                                    "disk": 40},
-               "8 GB Performance": {"id": "performance1-8", "ram": 8192, "vcpus": 8, "rxtx_factor": 1600,
-                                    "disk": 40}}
-
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a performance1 flavor.
-        """
-        return {
-            "class": "performance1",
-            "policy_class": "performance_flavor"
-        }
-
-
-class RackspacePerformance2Flavor(Flavor):
-    """
-    A Rackspace performance flavor object representation
-    """
-    flavors = {"15 GB Performance": {"id": "performance2-15", "ram": 15360, "vcpus": 4,
-                                     "rxtx_factor": 1250, "disk": 40},
-               "30 GB Performance": {"id": "performance2-30", "ram": 30720, "vcpus": 8,
-                                     "rxtx_factor": 2500, "disk": 40},
-               "60 GB Performance": {"id": "performance2-60", "ram": 61440, "vcpus": 16,
-                                     "rxtx_factor": 5000, "disk": 40},
-               "90 GB Performance": {"id": "performance2-90", "ram": 92160, "vcpus": 24,
-                                     "rxtx_factor": 7500, "disk": 40},
-               "120 GB Performance": {"id": "performance2-120", "ram": 122880, "vcpus": 32,
-                                      "rxtx_factor": 10000, "disk": 40}}
-
-    def extra_specs_json(self):
-        """
-        Create a JSON-serializable data structure describing
-        ``OS-FLV-WITH-EXT-SPECS:extra_specs`` for a performance 2 flavor.
-        """
-        return {
-            "class": "performance2",
-            "policy_class": "performance_flavor"
-        }
+        if region_name not in self.regional_collections:
+            self.regional_collections[region_name] = (
+                RegionalFlavorCollection(tenant_id=self.tenant_id, region_name=region_name,
+                                         clock=self.clock))
+        return self.regional_collections[region_name]
