+++ mimic/model/nova_objects.py
@@ -18,10 +18,7 @@ from mimic.util.helper import (
 )
 from mimic.canned_responses.mimic_presets import get_presets
 from mimic.model.glance_objects import Image, random_image_list
-from mimic.model.flavor_objects import (
-    Flavor, RackspaceStandardFlavor, RackspaceComputeFlavor, RackspaceMemoryFlavor,
-    RackspaceOnMetalFlavor, RackspaceIOFlavor, RackspaceGeneralFlavor,
-    RackspacePerformance1Flavor, RackspacePerformance2Flavor)
+
 from mimic.model.behaviors import (
     BehaviorRegistryCollection, EventDescription, Criterion, regexp_predicate
 )
@@ -653,7 +650,6 @@ def metadata_to_creation_behavior(metadata):
     ["tenant_id", "region_name", "clock",
      Attribute("servers", default_factory=list),
      Attribute("image_store", default_factory=list),
-     Attribute("flavors_store", default_factory=list),
      Attribute(
          "behavior_registry_collection",
          default_factory=lambda: BehaviorRegistryCollection())]
@@ -670,14 +666,6 @@ class RegionalServerCollection(object):
         for server in self.servers:
             if server.server_id == server_id and server.status != u"DELETED":
                 return server
-
-    def flavor_by_id(self, flavor_id):
-        """
-        Retrieve a :obj:`Flavor` object by its ID.
-        """
-        for flavor in self.flavors_store:
-            if flavor.flavor_id == flavor_id:
-                return flavor
 
     def request_creation(self, creation_http_request, creation_json,
                          absolutize_url):
@@ -987,47 +975,6 @@ class RegionalServerCollection(object):
         }
         return dumps(result)
 
-    def create_flavors_list(self, flavor_classes):
-        """
-        Generates the data for each flavor in each flavor class
-        """
-        for flavor_class in flavor_classes:
-            for flavor, flavor_spec in flavor_class.flavors.iteritems():
-                if not self.flavor_by_id(flavor_spec['id']):
-                    flavor_name = flavor
-                    flavor_id = flavor_spec['id']
-                    ram = flavor_spec['ram']
-                    vcpus = flavor_spec['vcpus']
-                    network = flavor_spec['rxtx_factor']
-                    disk = flavor_spec['disk']
-                    tenant_id = self.tenant_id
-                    flavor = flavor_class(flavor_id=flavor_id, tenant_id=tenant_id,
-                                          name=flavor_name, ram=ram, vcpus=vcpus,
-                                          rxtx=network, disk=disk)
-                    self.flavors_store.append(flavor)
-
-    def list_flavors(self, include_details, absolutize_url):
-        """
-        Return a list of flavors with details.
-        Creates a random list of flavors if flavors were not created.
-        """
-        flavors = [RackspaceStandardFlavor, RackspaceComputeFlavor, RackspacePerformance1Flavor,
-                   RackspaceOnMetalFlavor, RackspacePerformance2Flavor, RackspaceMemoryFlavor,
-                   RackspaceIOFlavor, RackspaceGeneralFlavor]
-        self.create_flavors_list(flavors)
-        flavors = []
-        for flavor in self.flavors_store:
-            if self.region_name != "IAD" and isinstance(flavor, RackspaceOnMetalFlavor):
-                continue
-            if include_details:
-                flavors.append(flavor.detailed_json(absolutize_url))
-            else:
-                flavors.append(flavor.brief_json(absolutize_url))
-
-            result = {"flavors": flavors}
-
-        return dumps(result)
-
     def get_image(self, http_get_request, image_id, absolutize_url):
         """
         Return a image object if one exists from the list `/images` api,
@@ -1046,25 +993,6 @@ class RegionalServerCollection(object):
                           distro='linux', tenant_id=self.tenant_id)
             self.image_store.append(image)
         return dumps({"image": image.get_server_image_details_json(absolutize_url)})
-
-    def get_flavor(self, http_get_request, flavor_id, absolutize_url):
-        """
-        Return a flavor object if one exists from the list `/flavors` api,
-        else creates and adds the flavor to the :obj: `flavors_store`.
-        If the `flavor_id` is listed in `mimic.canned_responses.mimic_presets`,
-        then will return 404.
-        """
-        if flavor_id in get_presets['servers']['invalid_flavor_ref']:
-            return dumps(not_found("The resource could not be found.",
-                                   http_get_request))
-        flavor = self.flavor_by_id(flavor_id)
-        if flavor is None:
-            flavor = Flavor(flavor_id=flavor_id,
-                            name=flavor_id + "Mimic Test Instance",
-                            ram=1, tenant_id=self.tenant_id, vcpus=2, rxtx=200, disk=3)
-
-            self.flavors_store.append(flavor)
-        return dumps({"flavor": flavor.detailed_json(absolutize_url)})
 
 
 @attributes(["tenant_id", "clock",
